test: Refactor out assertOutput
